# Vectorized implementation of reflex motion of the Earth correction